float to double

+++ include/point.h
@@ -1,7 +1,9 @@
 #ifndef POINT_H
 #define POINT_H
+
 typedef struct point
 {
-  float x = 0.f, y = 0.f;
+  double x = 0.f, y = 0.f;
 } Point;
+
 #endif
